Merge Pull Request
Try to keep the repository as simple as possible, so that any person accessing this repo would able to understand the whole project structure at a
glance. Also try to make your commits as small as possible. It helps to understand what the commits does and makes it easier to merge.